requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.4
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
email-validator==2.2.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
redis==6.4.0
rq==2.6.0
Pillow==11.3.0
httpx==0.28.1
